requirements.txt
PySide6
mysql-connector-python
pydantic-settings
PyMySQL
reportlab
openpyxl
matplotlib
SQLAlchemy